FSA: Remove disabling of BFCache on use of any local file
BFCache is currently disabled when using the File System Access API to
interact with _any_ local files in any way. This was added back in
https://crrev.com/c/1837813 due to its use of
RenderFrameHost::IsCurrent, which at the time returned true
for iframes

However, the behavior of RenderFrameHost::IsCurrent was changed in
https://crrev.com/c/2060374 such that it's no longer necessary to
disable BFCache in this specific case.
See https://crbug.com/1259861 for more details.

Note that BFCache is still disabled when permission to a file is
requested from the File System Access API.
This behavior was added in https://crrev.com/c/3217485 and requires
further work to address

Bug: 1241174, 1259861
Change-Id: Id0f00c9a801cb38331a4584682fd26bc18beab6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3169116
Commit-Queue: Austin Sullivan <asully@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143092}