[Passwords] Extract MetadataStoreStore as an inner class in LoginDB
Before this CL:
LoginDatabase acted as the sync MetadataStore for synced passwords.

After this CL:
The functionality is extracted to an inner class. This allows the
LoginDatabase to expose multiple sync MetadataStore's if multiple
one are needed.
More details are in the linked bug.

Follow-up CLs will make the PasswordSyncMetadataStore more generic.

This is a purely mechanical CL.

Bug: 1444700
Change-Id: I0468121eaa5108d8f182113ae4b49aa4d3a80858
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4521357
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1143132}