[ios][BVC] Refactor displayWebState into displayTabView
The main goal of this CL is to make method displayWebState independent
on webState. To achieve this the webState view will be used in that
method.

As a prerequisite note that displayWebState is always called on the
current webState, so there is no need to pass this information to the
method.

Some methods will change name:
 - displayWebState -> displayTabView
 - displayWebStateIfActive -> displayTabViewIfActive
 - sideSwipwRedisplayWebState -> sideSwipwRedisplayTabView
 - ntpFrameforWebState -> ntpFrameForCurrentWebState (also this method
   is always called with current webState, so no need to pass the
   webState to it).

In addition this CL creates a new method: updateWebStateVisibility
which will update the currentWebState visibility.

Bug: 1329088
Change-Id: I27725277804f9e097041b1c446cdc1615c25af38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4492775
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1143214}