Enable AutofillLeipzigFeatureCompatibilityLaunch on ToT
The feature has all necessary approval for launch (launch/4231286). As
per go/finch-best-practices, it is activated on ToT and the field trial
config entry is removed.

This breaks some PersonalDataManager tests related to the wallet address
conversion. This is because the address conversion is deprecated with
this launch. Hence, the tests are disabled, depending on the feature's
state. Again as per Finch best practices, the tests are not removed
entirely yet. This should only be done during clean up, once the feature
flag is removed.

Bug: 1348294
Change-Id: I0169ee2ff7f93cac8e3fc48e578678e1d563abec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520694
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1143254}