Allow programmatic clicks on AX nodes created from CSS content property
Previously, assistive tech trying to click an AX Node by calling
DoDefaultAction() would do nothing if the AX Node was created from a
layout tree traversal in AXNodeObject::AddChildrenImpl(). This is
because these AX nodes have no associated node or element in the DOM,
so the logic in AXObject::OnNativeClickAction doesn't do anything.
Now, this change attempts to fix this in two parts.

First, in AXObject::NativeClickAction, we will attempt to climb the
ancestor chain until we find a parent AXObject that has an element
if the current AXObject doesn't already have an element.

Second, we will give PseudoElements an AccessKeyAction() implementation
that just passes the request to the originating element. I felt weird
about this since the function name implies pseudo elements can have an
accesskey attribute which is only allowed on HTMLElements, but maybe
this is okay since SVGElements do the same thing?

Bug: 1174929
Change-Id: I49cfe9eb196078ac62a39ba4ab82dae97dc3189c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4007842
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143284}