Wallet Metadata: Document when local data can exist in MergeFullSyncData
For this data type, there usually shouldn't be any pre-existing local
data at the time MergeFullSyncData() gets called - but in some corner
cases, it may happen. This is covered by existing tests.
This CL adds an explanatory comment in MergeFullSyncData(), so that the
next reader doesn't have to wonder if this is dead code.

Fixed: 1377801
Change-Id: I3918363c280ee56c30981c058e52934a86211d98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518279
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143314}