[Clank] Fix Visibility with SpareTab
Currently the Visibility of the WebContents is updated when the Tab is
added to the TabStrip. But this is resulting in not recording the
PageLoadMetrics as the WebContents is in background.

This CL updates the WebContents Visibility before loading the URL to
match the flow without spare tab and trigger metrics logging as expected.

BUG=1441758

Change-Id: Ib069a09c22499d78ade1277fe714a117ab9cce5f
Validate-Test-Flakiness: skip
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4498087
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143318}