[iOS] Fix reading list sections order after item move
Previously, if there's one item in the reading list read section, and no
item in the unread section, and if the promo is shown, moving the item
from read to unread first moves the item above the promo, then moves it
below it quickly.

This CL fixes this buggy behavior.

Fixed: 1440158
Change-Id: Ie4b8a1816c6b6de38ab570e8f7dad186941eb901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4507471
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1143325}