Revert "Revert "[CrOS Cellular] Prevent cyclical shill calls to SetProperties().""
The test was consistently failing on Step "unit_tests on Ubuntu-18.04" failing on builder "chromium/ci/linux-cfm-rel" because DCHECK may not
be on, causing the mock expected call to be 1 instead of 2. See
crbug/1444602.

This CL reverts the commit, changes the DCHECK statement to CHECK,
which is always executed [1], and also changes some debug statements.

[1] https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:base/check.h;l=35-37;drc=68d2af39d3897338e231b24fb358dce0246c3ff5

This reverts commit 86ee3c141525f302c7efa399a05aef2549420572.

Fixed: b/280886640
Test: unit_test
Change-Id: I860b5f33c2da9d6aa9f75b8f434020c2577cf0d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518634
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143334}