Create ip_protection library in chrome/browser
This library will control the following:
- Getting an OAuth token for a logged in user.
- Using the OAuth token to make requests to a Google owned Auth server
- Receiving tokens from the request to the Auth server and sending them
back to the network service.

These tokens will be attached as a header to any request going to the
IP protection proxy servers.

I will follow with tests as this is not being called anywhere and
is just a skeleton with no complete flow to test.

Bug: 1444621
Change-Id: I482291e3898665eea11b7534f29691e8eacc08c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4519226
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Brianna Goldstein <brgoldstein@google.com>
Cr-Commit-Position: refs/heads/main@{#1143336}