FLEDGE: Alter how bidCurrency redaction for reportWin works.
Previously, we would pass the bid currency if it was configured in
perBuyerCurrency, including passing ??? if not specified; but as Paul
pointed out when reviewing the explainer changes, this lets
generateBid() leak a bit by omitting or including its currency tag.

So instead, just always pass the configured perBuyerCurrency, if any.

Change-Id: I9f8380d78cf4030b8fe713da73d8d74d6b081912
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4526537
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143338}