[omnibox][categorical] Adds omnibox::SuggestType to ACMatch
ACMatch::Type is currently inferred from the server reported string
value representation of the `omnibox::SuggestType` enum values.

The server reported types are subsequently ignored and just before the
navigation are inferred back from the ACMatch::Type to be reported in
the searchbox stats. This indirection adds complexity and is error
prone.

This change sets the omnibox::SuggestType on the ACMatch and populates
it at the provider level. This change does not introduce any behavior
changes, however, it adds a DCHECK to validate that the `suggest_type`
found on the ACMatch is the same as the one inferred from
ACMatch::Type. Once sure, the logic to infer the type can be removed
in favor of the suggest type found on the match.

This is being done to support categorical suggestions in Chrome.
Those suggestions have a different type (TYPE_CATEGORICAL_QUERY)
than the entity suggestions (TYPE_ENTITY) but come with an image
like the entity suggestions. In order to support these suggestions
with minimal changes in the code, they can be treated just like
`ACMatchType::SEARCH_SUGGEST_ENTITY` matches, however, they must
report the correct `TYPE_CATEGORICAL_QUERY` when selected. This
is difficult to accomplish right now due to type->match_type->type conversion and requires relying on the subtypes which does not
guarantee mutual exclusion like the type does. By reporting the
`omnibox::SuggestType` on the ACMatch directly, the correct behavior
will be automatically achieved.

Bug: 1444604
Change-Id: I4e1a2f2a136963b8afff55f25d53f3509145d5e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523376
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143415}