viz: Don't damage full surface during view transitions.
If the frame for a Surface was replaced by SurfaceAnimationManager, the
code assumes the whole surface needs to be damaged. This was needed when
the animation was viz driven, that code didn't compute the damage rect.
But now that its all driven in the renderer, CC should be setting up
the damage rects correctly.

R=kylechar@chromium.org

Bug: 1427721
Change-Id: Ibdd7d7e022c936c3fcf09f2b5719d994041e6636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523383
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143430}