personalization: time of day wallpaper first
When user is viewing available wallpaper collections, give the
time of day wallpaper collection the first spot, before local
images and google photos.

If the time of day collection is expected but does not exist,
will log an error and shift all collections one slot left.
If the time of day collection is not expected but does exist,
it is ignored and a warning is logged.

Bug: b:280467959
Test: browser_tests --gtest_filter=*PersonalizationAppComponent*
Change-Id: I284171d2cc1711892003e7191c623a10bcb02d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4525076
Reviewed-by: Jason Thai <jasontt@chromium.org>
Commit-Queue: Jeffrey Young <cowmoo@google.com>
Cr-Commit-Position: refs/heads/main@{#1143442}