Don't close dialogs in response to cancel event
A *trusted* cancel event can close dialog elements. This is hooked up to
the default keyboard event handler, which fires a trusted cancel event,
in order to make the escape key close dialog elements. When the cancel
event has preventDefault called on it, it prevents the dialog element
from being closed because the dialog's DefaultEventHandler won't be
called anymore.

This caused an issue where a new trusted cancel event started being
emitted by file input elements, which unexpectedly caused the dialog to
be closed.

This patch replaces the dialog's DefaultEventHandler with a check to see
if the event was defaultPrevented where it gets emitted in the keyboard
event handler.

Fixed: 1442824
Change-Id: Icce30b3945dd900d90758b4ea3f223008ab0ccd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4527455
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143502}