Ignore extra clipboard selection reads
We are seeing some crashes on wayland with the error:
    GDBus.Error:org.freedesktop.DBus.Error.LimitsExceeded: Tried to read in parallel

It looks like this is happening when multiple SelectionReads happen.

There are a few different solutions to this, but this CL will make it so that any extra requests are ignored.  If any requests get eaten in this way, then we will queue another read so that we will still have the latest clipboard contents.

Change-Id: Icaf90a4f3c38b5580ca35f8ebed4f0df16b5fa31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4499088
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Max Booth <maxbooth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143519}