[A11y] Small simplification to ignored computation for block flows
This continues the work from CL:4194697 to simplify our code that decides whether or not a block flow is ignored. By simplifying, we improve performance and predictability, and have fewer cases where we miss an invalidation when one of the dependent conditions changes.

In this case, we remove code that checks to see if a cursor can be set inside of the block flow. As a result, a few more harmless generic containers are introduced to the hierarchy. It is not necessary to remove all uninteresting generic containers, and it is only an optimization to do so. In this case, the extra "magic" regarding NGCursor is not worth having a few extra generic containers in the tree.

The code is now clearer and less likely to run into a state where there is a stale tree structure containing a node with the wrong ignored state, confusing the serializer.

Bug: none
Change-Id: Id19f911547c56e4518d43e71e04fc25a04fc36b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4261382
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143545}