[iOS] Always use Password Bottom Sheet when available
This CL removed the previous logic which had been added in:
https://chromium-review.googlesource.com/c/chromium/src/+/4507791
which disabled the Password Bottom Sheet when the Keyboard
Accessory was shown.

This CL reverts that and instead always opens the bottom sheet
when possible, even if a password related field is already in
focus, except in the specific case where the user has already
started to enter data in that field.

Bug: 1440704, 1440734, 1444068
Change-Id: Iee2066e8cad969136cce6be380b69180cff4daf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518739
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143584}