[Extensions] Use browser context unique id in extension utils
Previously, extensions util generated a "unique id" given a browser
context and stored them in a `std::map<content::BrowserContext*, int>`. This was problematic because if the BrowserContext gets
destroyed, another BrowserContext could be placed in the same address
and get the "unique id" of the destroyed BrowserContext.

Instead, we can use the BrowserContext::UniqueId that guarantees a
unique id. We maintain the map generated by extension utils because
the types are different (int vs string) and because it's also used by
the renderer.

Bug: 1444279
Change-Id: If1e5a74d4652695133d6ffc15c869347dbbf9ebe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523080
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143597}