[PDF OCR] Disable the More Actions entry point
As per request from the accessibility analyst, disable the PDF OCR entry
point in the More Actions menu for accessibility consistency even if
the PDF OCR feature is enabled.

As newly added TODO comments indicate, this entry point in the More
Actions and its browser tests would be re-enabled when PDF OCR is
integrated with Select-to-Speak, which would begin in 2023 Q3.

AX-Relnotes: n/a.
Bug: 1393069
Change-Id: I788e7368b5772aef85ddf82c1200d64cf094dc7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523387
Reviewed-by: K. Moon <kmoon@chromium.org>
Commit-Queue: Kyungjun Lee <kyungjunlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1143602}