media/gpu/v4l2: Update Max Device Number for Kukui Bug
The ioctl shim was previously setup with a maximum device number of 20
for queries. This fails on kukui devices as the device has a number
greater than 20.

This updates the maximum device number to 40 as the path in /dev/video##
goes beyond 20.

This update results in H.264 stateless tests passing

BUG=b:282219255
TEST=tast -verbose run dut video.PlatformDecoding.v4l2_stateless_h264_baseline

Change-Id: I43182b79277a0dfa677480aebb4bd5ba7a88d685
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4528966
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143618}